requirements.txt
django~=3.1.2
djangorestframework~=3.13.1
psycopg2-binary

django-filter~=21.1